fix(deps): keep HA back-compat to 2023.10.0, scope homeassistant fixtures out of root conftest
Two problems surfaced after the homeassistant dev-dependency bump:

1. HouseholdBinCoordinator unconditionally passed config_entry= to
   DataUpdateCoordinator.__init__, but that parameter only exists on Home
   Assistant >=2024.11.0 - older cores raise TypeError on the unrecognized
   kwarg. This integration documents support back to 2023.10.0 (see
   COMPATIBILITY.md), so the previous change would have broken setup for
   anyone still on an HA core between 2023.10 and 2024.10. Now checked once
   via inspect.signature and only passed through when the installed core
   actually supports it - real end users are unaffected either way, since
   homeassistant itself is never part of what they install (manifest.json
   only requires uk-bin-collection; they always run against their own HA).

2. Root conftest.py imported homeassistant at module level, which CI's
   council-scraper BDD suite (uk_bin_collection/tests/step_defs/) also
   loads regardless of whether homeassistant is installed for that Python
   version - it broke that suite on CI with
   "ModuleNotFoundError: No module named 'homeassistant'". Moved the
   Home-Assistant-specific fixtures (hass, enable_custom_integrations, the
   frame-helper setup) into their own
   custom_components/uk_bin_collection/tests/conftest.py, which pytest only
   loads for that test tree.

Also documents the Python 3.14 requirement for running the full dev test
suite in CONTRIBUTING.md and the release-workflow docs.
